+++ app/src/main/java/com/example/cardiocare/Splash.java
@@ -1,0 +1,32 @@
+package com.example.cardiocare;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.view.Window;
+import android.view.WindowManager;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+
+public class Splash extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        requestWindowFeature(Window.FEATURE_NO_TITLE);
+        //HIDE ACTION BAR
+        getSupportActionBar().hide();
+        //HIDE TITLE BAR
+        getWindow().setStatusBarColor(ContextCompat.getColor(Splash.this,R.color.splash0));
+       // getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
+        setContentView(R.layout.activity_splash);
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                startActivity(new Intent(Splash.this,MainActivity.class));
+                finish();
+            }
+        },3000);
+    }
+}
